Renaming entities and folders.
