+++ TOT.Data/TOTUnitOfWork.cs
@@ -1,6 +1,6 @@
 ﻿using System.Threading.Tasks;
 using TOT.Data.Repositories;
-using TOT.Entities.Request_Entities;
+using TOT.Entities.TimeOffRequests;
 using TOT.Interfaces;
 
 namespace TOT.Data
