--- TOT.Entities/Policy_Entities/TimeOffPolicyCheckers.cs
@@ -1,9 +1,0 @@
-﻿namespace TOT.Entities.Policy_Entities
-{
-    public class TimeOffPolicyCheckers
-    {
-        public int Id { get; set; }
-        public Positions Position { get; set; }
-        public int Amount { get; set; } 
-    }
-}
