--- TOT.Entities/Request_Entities/RequestStatus.cs
@@ -1,8 +1,0 @@
-﻿namespace TOT.Entities.Request_Entities
-{
-    public class RequestStatus
-    {
-        public int Id { get; set; }
-        public string Title { get; set; }
-    }
-}
