--- TOT.Entities/Request_Entities/TimeOffType.cs
@@ -1,8 +1,0 @@
-﻿namespace TOT.Entities.Request_Entities
-{
-    public class TimeOffType
-    {
-        public int Id { get; set; }
-        public string Title { get; set; }
-    }
-}
