+++ TOT.Entities/TimeOffPolicies/AccrualSchedule.cs
@@ -1,4 +1,4 @@
-﻿namespace TOT.Entities.Policy_Entities
+﻿namespace TOT.Entities.TimeOffPolicies
 {
     public class AccrualSchedule
     {
