+++ TOT.Entities/TimeOffPolicies/TimeOffPolicyApproval.cs
@@ -1,0 +1,9 @@
+﻿namespace TOT.Entities.TimeOffPolicies
+{
+    public class TimeOffPolicyApproval
+    {
+        public int Id { get; set; }
+        public EmployeePosition Position { get; set; }
+        public int Amount { get; set; } 
+    }
+}
