+++ TOT.Entities/TimeOffRequests/TimeOffType.cs
@@ -1,0 +1,7 @@
+﻿namespace TOT.Entities.TimeOffRequests
+{
+    public enum TimeOffType : int
+    {
+        PayedTimeOff = 1
+    }
+}
